Remove circular deps by folding `default-scope` into `scope`.
